+++ www/whatsnew.md
@@ -6,6 +6,7 @@ permalink: /whats-new/
 
 ### Upcoming Changes
  - Xii'da the Curious now correctly doubles the minions as well as their openers. (1544)
+ - Exchange Wares now actually does its effect. (1544)
 
 ### 0.8.66-3.0.9 (February 6, 2020)